+++ src/Form/AccountType.php
@@ -30,24 +30,6 @@ class AccountType extends AbstractType
                 'help' => 'create_account_form.account_value_help',
                 'required' => true,
                 'row_attr' => ['class' => 'form-row'],
-            ])
-            ->add('status', TextType::class, [
-                'label' => 'create_account_form.account_status_label',
-                'label_attr' => ['class' => 'label'],
-                'attr' => ['class' => 'form-element'],
-                'help_attr' => ['class' => 'form-help'],
-                'help' => 'create_account_form.account_status_help',
-                'required' => true,
-                'row_attr' => ['class' => 'form-row'],
-            ])
-            ->add('category', TextType::class, [
-                'label' => 'create_account_form.account_category_label',
-                'label_attr' => ['class' => 'label'],
-                'attr' => ['class' => 'form-element'],
-                'help_attr' => ['class' => 'form-help'],
-                'help' => 'create_account_form.account_category_help',
-                'required' => true,
-                'row_attr' => ['class' => 'form-row'],
             ]);
     }