+++ CHANGELOG.md
@@ -8,6 +8,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+* Removed usused fields from account and client.
 * Added view delete protection
 * Adds views filtering.
 * Adds user administration.